feat(*): move plugin to lib folder, tests to test folder
