+++ jest.setup.js
@@ -1,3 +1,5 @@
-const getTestRule = require('jest-preset-stylelint/getTestRule');
+const { getTestRule, getTestRuleConfigs } = require('jest-preset-stylelint');
+const plugins = [require('./')];
 
-global.testRule = getTestRule({ plugins: ['./'] });
+global.testRule = getTestRule({ plugins });
+global.testRuleConfigs = getTestRuleConfigs({ plugins });
